+++ devices/cpu/cpu_engine.h
@@ -23,12 +23,12 @@ OIDN_NAMESPACE_BEGIN
     int getNumThreads() const { return device->numThreads; }
 
     // Ops
-    bool isConvSupported(Fusion fusion) override;
-    bool isConcatConvSupported(Fusion fusion) override;
+    bool isConvSupported(Fusion fusion) const override;
+    bool isConcatConvSupported(Fusion fusion) const override;
   #if !defined(OIDN_BNNS)
     Ref<Conv> newConv(const ConvDesc& desc) override;
-    Ref<ConcatConv> newConcatConv(const ConcatConvDesc& desc) override;
   #endif
+    Ref<ConcatConv> newConcatConv(const ConcatConvDesc& desc) override;
     Ref<Pool> newPool(const PoolDesc& desc) override;
     Ref<Upsample> newUpsample(const UpsampleDesc& desc) override;
     Ref<Autoexposure> newAutoexposure(const ImageDesc& srcDesc) override;