+++ devices/metal/metal_engine.h
@@ -49,8 +49,10 @@ OIDN_NAMESPACE_BEGIN
     Ref<Tensor> newTensor(const Ref<Buffer>& buffer, const TensorDesc& desc, size_t byteOffset) override;
 
     // Ops
-    bool isConvSupported(Fusion fusion) override;
+    bool isConvSupported(Fusion fusion) const override;
+    bool isConcatConv2Supported(Fusion fusion) const override;
     Ref<Conv> newConv(const ConvDesc& desc) override;
+    Ref<ConcatConv2> newConcatConv2(const ConcatConvDesc& desc) override;
     Ref<Pool> newPool(const PoolDesc& desc) override;
     Ref<Upsample> newUpsample(const UpsampleDesc& desc) override;
     Ref<Autoexposure> newAutoexposure(const ImageDesc& srcDesc) override;