+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Version History
 
 -   Improve performance and reduce memory usage for Intel GPUs with XMX support
     and CPUs with AMX-FP16 support
+-   Fixed corrupted output on Apple M5 Pro/Max GPUs using Metal
 
 ### Changes in v2.4.1: